pyproject.toml
[project]
name = "spindle-ai"
version = "0.1.0"
description = "Pull-based agent framework where you control the loop"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
gemini = [
    "google-genai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]
all = [
    "spindle-ai[gemini,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spindle"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring real LLM API access",
]

[tool.ruff]
target-version = "py311"
line-length = 100
